+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ## Update
+## [0.0.4] - 2020-04-12
+* Endpoint for create user and login added.
+## Update
+## [0.0.3] - 2020-04-12
+* Logger for js added.
+## [0.0.2] - 2020-04-12
+* ESLint added.
+* Express Basic Auth added.
+## Update
 ## [0.0.1] - 2020-04-11
 * Initial project changelog and readme added.